Minor fixes
y a z a d o l b a l s y a p i s a t ' e t o t s h * t

+++ supplyGood/ViewSupply.cs
@@ -38,7 +38,7 @@ namespace supplyGood
                     }
                     catch (Exception ex) { }
                 }
-                return Math.Round((double)d, 2).ToString();
+                return Math.Round(d, 2).ToString();
             }
         }
         int GetIDByAllAttributes
